requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
asn1crypto==1.5.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.82
botocore==1.42.82
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
distro==1.9.0
filelock==3.25.2
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.14
langchain-core==1.2.25
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.4
playwright==1.58.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.1
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
snowflake-connector-python==4.4.0
sortedcontainers==2.4.0
soupsieve==2.8.3
streamlit==1.56.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
